+++ auto_route/lib/src/router/parser/route_information_parser.dart
@@ -17,12 +17,30 @@ class DefaultRouteParser extends RouteInformationParser<UrlState> {
   /// Passed to [RouteMatcher.matchUri]
   final bool includePrefixMatches;
 
+  /// Clients can use this to intercept deep-links
+  /// coming from platform and transform the [Uri].
+  /// Especially useful to remove pathPrefixes use in filter-intent on Android
+  /// or appLinks on Apple
+  ///
+  /// It's recommended to return [SynchronousFuture] of the transformed [Uri]
+  /// if you're doing any async operation in the transformer
+  final DeepLinkTransformer? deepLinkTransformer;
+
   /// Default constructor
-  DefaultRouteParser(this._matcher, {this.includePrefixMatches = false});
+  DefaultRouteParser(
+    this._matcher, {
+    this.includePrefixMatches = false,
+    this.deepLinkTransformer,
+  });
 
   @override
-  Future<UrlState> parseRouteInformation(RouteInformation routeInformation) {
-    final resolvedUri = _normalize(routeInformation.uri);
+  Future<UrlState> parseRouteInformation(
+    RouteInformation routeInformation,
+  ) async {
+    final resolvedUri = _normalize(
+      await deepLinkTransformer?.call((routeInformation.uri)) ??
+          routeInformation.uri,
+    );
     var matches = _matcher.matchUri(resolvedUri,
         includePrefixMatches: includePrefixMatches);
     return SynchronousFuture<UrlState>(