+++ auto_route/lib/src/router/widgets/auto_route_navigator.dart
@@ -63,7 +63,8 @@ class AutoRouteNavigatorState extends State<AutoRouteNavigator> {
 
   void _updateDeclarativeRoutes() {
     final delegate = AutoRouterDelegate.of(context);
-    var newRoutes = widget.declarativeRoutesBuilder!(widget.router.pendingRoutesHandler);
+    var newRoutes =
+        widget.declarativeRoutesBuilder!(widget.router.pendingRoutesHandler);
     if (!const ListEquality().equals(newRoutes, _routesSnapshot)) {
       _routesSnapshot = newRoutes;
       widget.router.updateDeclarativeRoutes(newRoutes);
@@ -86,8 +87,12 @@ class AutoRouteNavigatorState extends State<AutoRouteNavigator> {
     final navigator = widget.router.hasEntries
         ? Navigator(
             key: widget.router.navigatorKey,
-            observers: [widget.router.pagelessRoutesObserver, ...widget.navigatorObservers],
-            restorationScopeId: widget.navRestorationScopeId ?? widget.router.routeData.restorationId,
+            observers: [
+              widget.router.pagelessRoutesObserver,
+              ...widget.navigatorObservers
+            ],
+            restorationScopeId: widget.navRestorationScopeId ??
+                widget.router.routeData.restorationId,
             pages: widget.router.stack,
             onPopPage: (route, result) {
               if (!route.didPop(result)) {