+++ auto_route/CHANGELOG.md
@@ -1,15 +1,35 @@
-## 8.0.3
+## 8.1.2
 
- - **FIX**: issue related to casting Custom Routes.
+- **FIX**: Fixed back gesture when nested navigator only has one entry #1940
+
+
+## 8.1.1
+
+- **FIX**: Can not swipe back on iOS when using nested navigation #1932.
+
+
+## 8.1.0
+
+- **FEAT**: add deep-link transformer.  Allows parsing the URI for deep links before the
+  route matcher performs a match on the URI.  Especially useful when you use intent-filter
+  prefix on Android to only handle links from e.g com.example/FILTER.
+
+  ```dart
+  RouterConfig `<UrlState>` config({
+  deepLinkTransformer : Deelkink.prefixStripper('FILTER'),
+  )
+  ```
+
+  The resulting URI will be : com.example
 
 ## 8.0.2
 
- - **FIX**: only disable parent back gesture if the active child can pop.
- - **FIX**: AutoTabsRouterTabBarState was disposed with active Ticker #1910
+- **FIX**: only disable parent back gesture if the active child can pop.
+- **FIX**: AutoTabsRouterTabBarState was disposed with active Ticker #1910
 
 ## 8.0.1
 
- - **FIX**: back-gesture does not respect sub-routes.
+- **FIX**: back-gesture does not respect sub-routes.
 
 ## 8.0.0
 
@@ -18,7 +38,7 @@
 - **BREAKING CHANGE**: deep-links with no host e.g scheme://page/sub-page are no longer
   automatically handled, flutter's deep-link handler will treat [page] as host so we end up
   with [/sub-page] as a path which of course not going to match. if your deep-links are hostless
-  you'll need to handle them manually inside deepLinkBuilder.
+  you'll need to handle them manually inside deepLinkTransformer or deepLinkBuilder.
 
 ## 7.9.2
 
@@ -246,7 +266,7 @@ the [Migration guide](https://github.com/Milad-Akarie/auto_route_library/tree/v6
 - Add title builder for AutoRoute(title: (ctx,data){})
 - Docs are changed to reflect the new changes so make sure you re-read them.
 
- ---
+---
 
 ## [5.0.4]
 
@@ -467,7 +487,7 @@ the [Migration guide](https://github.com/Milad-Akarie/auto_route_library/tree/v6
 
 ## [2.1.0] Breaking changes!
 
-- AutoRouteGuard no longer returns a future<bool> and passes in a resolver. [Breaking]
+- AutoRouteGuard no longer returns a future`<bool>` and passes in a resolver. [Breaking]
 - Remove onInitialRoutes from AutoRouterDelegate. [Breaking]
 - Rename RoutingControllerScope to RouterScope. [Breaking]
 - Expose provided navigatorObservers in RouterScope
@@ -602,7 +622,7 @@ the [Migration guide](https://github.com/Milad-Akarie/auto_route_library/tree/v6
 
 - Rebuild auto_route to work with Navigator 2.0
 
------------------------------------------------
+---
 
 ## [0.6.9]
 
@@ -693,7 +713,7 @@ the [Migration guide](https://github.com/Milad-Akarie/auto_route_library/tree/v6
 
 - Add ability to pass initial route arguments to ExtendedNavigator
 - Change single route parameters will have argument holder classes too as requested
-- Fix ExtendedNavigator.ofRouter<T>() returns null in inspector mode
+- Fix ExtendedNavigator.ofRouter`<T>`() returns null in inspector mode
 
 ## [0.4.2]